Create a constants file to centralize their usage
