+++ lib/utils/constants.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/painting.dart';
+import 'package:flutter/material.dart';
+
+const PADDING = 5.0;
+const CARD_TITLE_STYLE = TextStyle(
+    color: Color(0xff40B491), fontSize: 20.0, fontWeight: FontWeight.w500);
+const CARD_TEXT_STYLE =
+    TextStyle(color: Colors.black, fontSize: 16.0, fontWeight: FontWeight.w400);
+const RESULT_TEXT_STYLE =
+    TextStyle(color: Colors.black, fontSize: 18.0, fontWeight: FontWeight.w400);
+const CARD_ICON_COLOR = Colors.black;
+const CARD_ICON_TEXT_SPACE = 10.0;
+const CARD_BUTTON_TEXT_STYLE = TextStyle(
+    color: Color(0xff40B491), fontSize: 16.0, fontWeight: FontWeight.w500);
+const CARD_ELEVATION = 10.0;
